Apply role.upgrader to relevant creep
In our main module, all creeps run the same role. We need to divide their behavior depending on the previously defined property Creep.memory.role by connecting the new module.

Apply the logic from the module role.upgrader to the creep with the role upgrader and check how it performed.