Usage of StructureSpawn.renewCreep
Now the memory of the deceased is relegated to oblivion which saves us resources.

Apart from creating new creeps after the death of old ones, there is another way to maintain the needed number of creeps: the method StructureSpawn.renewCreep. Creep aging is disabled in the Tutorial, so we recommend that you familiarize yourself with it on your own.

Documentation:
StructureSpawn.renewCreep