Set the harvester role to bring energy to the tower
A tower uses energy, so let’s set the harvester role to bring energy to the tower along with other structures. To do this, you need to add the constant STRUCTURE_TOWER to the filter of structures your harvester is aimed at.

Add STRUCTURE_TOWER to the module role.harvester and wait for the energy to appear in the tower